ingestion-service/requirements.txt
python-dotenv
boto3

black
ruff